Fix Minor 2 and Minor 3 from opus-hardening-review
Minor 3 (layering): Move user_has_branch_access() from ury_yield_variance.py
to ury/report_api/utils.py alongside require_manager(), the established
shared-utils home for cross-cutting permission helpers. This eliminates the
cross-module import smell where ury_issue_authorization was importing from
ury_yield_variance, and centralizes the auth helper with its semantic peer.

Minor 2 (correctness hardening): Add parentfield: "user" to the
user_has_branch_access() query filter. The fieldname is "user" per
fixtures/custom_field.json's Branch-user entry. This prevents a false
positive in the future if a second parent table is added to URY User.

Update both call sites to import from the new location:
- ury_yield_variance.py: import user_has_branch_access from utils
- ury_issue_authorization.py: import user_has_branch_access from utils
  instead of from ury_yield_variance

Track-Item: D3
